[feat] 작업부 Nested DTO화
