--- src/main/kotlin/com/chatandpay/api/dto/OtherBankAccountListResponseDTO.kt
@@ -1,11 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.common.ApiResponse
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class OtherBankAccountListResponseDTO (
-    val accountList: List<OtherBankAccountRawResponseDTO>
-): ApiResponse()
