+++ src/main/kotlin/com/chatandpay/api/service/AccountService.kt
@@ -1,8 +1,7 @@
 package com.chatandpay.api.service
 
 import com.chatandpay.api.domain.OtherBankAccount
-import com.chatandpay.api.dto.OtherBankAccountListResponseDTO
-import com.chatandpay.api.dto.OtherBankAccountRawResponseDTO
+import com.chatandpay.api.dto.AccountDTO
 import com.chatandpay.api.dto.OpenApiDTO
 import com.chatandpay.api.exception.RestApiException
 import com.chatandpay.api.repository.AccountRepository
@@ -32,13 +31,13 @@ class AccountService (
 
     }
 
-    fun getExternalAccount(userId: Long): OtherBankAccountListResponseDTO {
+    fun getExternalAccount(userId: Long): AccountDTO.OtherBankAccountListResponseDTO {
 
         val accountList = accountRepository.findByPayUserId(userId).map {
-            OtherBankAccountRawResponseDTO(it.bankCode, it.accountNumber, it.accountName, it.autoDebitAgree)
+            AccountDTO.OtherBankAccountRawResponseDTO(it.bankCode, it.accountNumber, it.accountName, it.autoDebitAgree)
         }
 
-        return OtherBankAccountListResponseDTO(accountList)
+        return AccountDTO.OtherBankAccountListResponseDTO(accountList)
 
     }
